requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis & Caching
redis==5.0.1
celery==5.3.4

# Vector Database
qdrant-client==1.7.0
sentence-transformers==2.2.2

# Knowledge Graph
neo4j==5.15.0

# Google Services
google-api-python-client==2.108.0
google-auth==2.25.2
# NEW: Crucial for user-based OAuth flow
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0

# LLM & AI
google-generativeai==0.3.2
langchain==0.0.350
langchain-google-genai==0.0.6

# Document Processing
python-multipart==0.0.6
python-docx==1.1.0
python-pptx==0.6.23
PyPDF2==3.0.1
pymupdf==1.23.8
openpyxl==3.1.2
pandas==2.1.4
unstructured==0.11.8

# Web & HTTP
httpx==0.25.2
websockets==12.0
aiofiles==23.2.1

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
rich==13.7.0
typer==0.9.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
